fix(DPAV-2060): fix execute_with_timeout reset (#155)
This was trying to use the transaction again when it is in a failed state
by issueing statement_timeout SET in finally block.

Instead use LOCAL so it only applies to the transaction, and reset it
after the query on success only. If it timesout or fails it will be
reset when the transaction ends anyway so no need for us to set it back.

Also increased the default timeout and removed redundant fallback value
that was added for testing config issue previously.